🗺️ Atlas: [architectural change] Splitting Blob modules into submodules
Refactored `src/storage/migration.rs` and `src/index/vector/distributed.rs` into submodules (`mod.rs` and `tests.rs`) to address the Blob anti-pattern and improve navigability. Recorded findings in `.jules/atlas.md`.
