+++ .jules/atlas.md
@@ -90,3 +90,10 @@
 ## 2026-06-21 - Splitting Redb Cold Storage Blob
 **Tangle:** `src/storage/redb_cold_storage.rs` was over 4100 lines long, a "Blob" module holding both core cold storage logic and over 2000 lines of tests. It made navigation and understanding the core operations difficult.
 **Blueprint:** Refactored into a `src/storage/redb_cold_storage/` module. Kept the core data definitions and implementation in `mod.rs` (reduced to ~2000 lines) and moved all tests to `tests.rs` (~2000 lines), maintaining the `#[cfg(test)]` block functionality but with better physical separation.
+## 2026-06-21 - Splitting Migration Blob
+**Tangle:** `src/storage/migration.rs` was over 3000 lines long, a "Blob" module holding both core migration logic and nearly 1000 lines of tests. It made navigation and understanding the core operations difficult.
+**Blueprint:** Refactored into a `src/storage/migration/` module. Kept the core data definitions and implementation in `mod.rs` and moved all tests to `tests.rs`, maintaining the `#[cfg(test)]` block functionality but with better physical separation.
+
+## 2026-06-21 - Splitting Distributed Vector Index Blob
+**Tangle:** `src/index/vector/distributed.rs` was over 2000 lines long, a "Blob" module holding both core distributed vector index logic and over 1000 lines of tests.
+**Blueprint:** Refactored into a `src/index/vector/distributed/` module. Kept the core data definitions and implementation in `mod.rs` and moved all tests to `tests.rs`, maintaining the `#[cfg(test)]` block functionality but with better physical separation.
